app/requirements.txt
psycopg2-binary==2.9.3 
Django==4.0.1
djangorestframework==3.13.1
django-phonenumber-field[phonenumbers]==6.0.0
phonenumbers==8.12.41
django_filter==21.1
Pillow==9.0.0
gunicorn==20.1.0